+++ packages/iov-jsonrpc/src/id.spec.ts
@@ -1,13 +1,21 @@
-import { makeId } from "./id";
+import { makeJsonRpcId } from "./id";
 
 describe("id", () => {
-  describe("makeId", () => {
-    it("works", () => {
-      const firstId = makeId();
-      expect(firstId).toBeGreaterThanOrEqual(1);
-      expect(makeId()).toEqual(firstId + 1);
-      expect(makeId()).toEqual(firstId + 2);
-      expect(makeId()).toEqual(firstId + 3);
+  describe("makeJsonRpcId", () => {
+    it("returns a string or number", () => {
+      const id = makeJsonRpcId();
+      expect(["string", "number"]).toContain(typeof id);
+    });
+
+    it("returns unique values", () => {
+      const ids = new Set([
+        makeJsonRpcId(),
+        makeJsonRpcId(),
+        makeJsonRpcId(),
+        makeJsonRpcId(),
+        makeJsonRpcId(),
+      ]);
+      expect(ids.size).toEqual(5);
     });
   });
 });