+++ packages/iov-jsonrpc/src/index.ts
@@ -1,4 +1,4 @@
-export { makeId } from "./id";
+export { makeJsonRpcId } from "./id";
 export { JsonRpcClient, SimpleMessagingConnection } from "./jsonrpcclient";
 export {
   JsonCompatibleArray,