+++ packages/iov-jsonrpc/src/jsoncompatibledictionary.spec.ts
@@ -1,4 +1,3 @@
-import { makeId } from "./id";
 import {
   isJsonCompatibleArray,
   isJsonCompatibleDictionary,