+++ packages/iov-jsonrpc/types/id.d.ts
@@ -1,1 +1,8 @@
-export declare function makeId(): number;
+/**
+ * Creates a new ID to be used for creating a JSON-RPC request.
+ *
+ * Multiple calls of this produce unique values.
+ *
+ * The output may be any value compatible to JSON-RPC request IDs with an undefined output format and generation logic.
+ */
+export declare function makeJsonRpcId(): number;