+++ packages/iov-jsonrpc/types/index.d.ts
@@ -1,4 +1,4 @@
-export { makeId } from "./id";
+export { makeJsonRpcId } from "./id";
 export { JsonRpcClient, SimpleMessagingConnection } from "./jsonrpcclient";
 export { JsonCompatibleArray, JsonCompatibleDictionary, JsonCompatibleValue, isJsonCompatibleArray, isJsonCompatibleDictionary, isJsonCompatibleValue, } from "./jsoncompatibledictionary";
 export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse, parseJsonRpcResponse2, parseJsonRpcError, parseJsonRpcErrorResponse, parseJsonRpcSuccessResponse, } from "./parse";